requirements.txt
streamlit
deepface
opencv-python-headless
Pillow
mtcnn
numpy
tensorflow
